Moves shared tag files into new folder
