+++ libs/api/database/src/lib/content/schemas/tags.schema.ts
@@ -1,6 +1,6 @@
 import { Schema, Prop, SchemaFactory } from '@nestjs/mongoose';
 import { Document, Schema as MongooseSchema } from 'mongoose';
-import { TagsModel, TagKind } from '@dragonfish/shared/models/content';
+import { TagsModel, TagKind } from '@dragonfish/shared/models/content/tags';
 import { nanoid } from 'nanoid';
 
 /**
