+++ libs/client/dashboard/src/lib/pages/tags-management/tags-management.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { TagsQuery, TagsService } from '@dragonfish/client/repository/dashboard/tags';
-import { TagKind, TagsModel } from '@dragonfish/shared/models/content';
+import { TagKind, TagsModel } from '@dragonfish/shared/models/content/tags';
 import { MatDialog } from '@angular/material/dialog';
 import { TagFormComponent } from '../../components/tags-management';
 import { PopupModel } from '@dragonfish/shared/models/util';
