+++ libs/client/repository/src/lib/dashboard/tags/tags.state.ts
@@ -1,4 +1,4 @@
 import { ActiveState, EntityState } from '@datorama/akita';
-import { TagsTree } from '@dragonfish/shared/models/content/tags.model';
+import { TagsTree } from '@dragonfish/shared/models/content/tags/tags.model';
 
 export interface TagsState extends EntityState<TagsTree, string>, ActiveState {}
