+++ libs/shared/src/lib/models/content/content.model.ts
@@ -2,7 +2,7 @@ import { UserInfo } from '../users';
 import { PubStatus } from './pub-status.enum';
 import { ContentKind } from './content-kind.enum';
 import { ContentRating } from '../works';
-import { TagsModel } from './tags.model';
+import { TagsModel } from '@dragonfish/shared/models/content/tags/tags.model';
 
 export interface ContentModel {
     readonly _id: string;
