+++ libs/shared/src/lib/models/content/tags/index.ts
@@ -1,0 +1,3 @@
+export type { TagsModel } from './tags.model';
+export type { TagsForm } from './tags-form.model';
+export { TagKind } from './tag-kind.enum';
